+++ R/par_angle_manual.R
@@ -1,26 +1,36 @@
-#' Calculate parallactic angles from manually entered data
+#' Calculate parallactic angles from Alpha, Delta and Stime for BTA location
 #'
-#' @return Parallactic angle in degrees from manually entered data.
+#' @param Alpha a character string (hh:mm:ss)
+#' @param Delta a character string (dd:mm:ss)
+#' @param Stime a character string (hh:mm:ss)
+#' @return Parallactic angle in degrees.
 #' @examples
+#' par_angle_manual(Alpha = '00:33:32.7', Delta = '55:18:59.5', Stime = '06:53:00')
 #' par_angle_manual()
 #' @export
-par_angle_manual <- function() {
+par_angle_manual <- function(Alpha = NULL, Delta = NULL, Stime = NULL) {
 
-  # BTA_latitude_rad <- 0.761894396137 # +43 deg 39' 12''
-  BTA_latitude_rad <- 43.6467 * pi / 180 # Dyachenko values
-  Alpha <- readline(prompt = 'Enter Alpha (hh:mm:ss): ')
+  BTA_latitude_rad <- 43.6467 * pi / 180 # BTA latitude in rad
+
+  if(is.null(Alpha)) {
+    Alpha <- readline(prompt = 'Enter Alpha (hh:mm:ss): ')
+  }
   Alpha_rad <- celestial::hms2deg(Alpha) * pi / 180
 
-  Delta <- readline(prompt = 'Enter Delta (dd:mm:ss): ')
+  if(is.null(Delta)) {
+    Delta <- readline(prompt = 'Enter Delta (dd:mm:ss): ')
+  }
   Delta_rad <- celestial::dms2deg(Delta) * pi / 180
 
-  Stime <- readline(prompt = 'Enter Stime (hh:mm:ss): ')
+  if(is.null(Stime)) {
+    Stime <- readline(prompt = 'Enter Stime (hh:mm:ss): ')
+  }
   Stime_rad <- celestial::hms2deg(Stime) * pi / 180
 
   Hour_angle_rad = Stime_rad - Alpha_rad
 
   Q_degr = atan(sin(Hour_angle_rad) / (tan(BTA_latitude_rad) * cos(Delta_rad) -
-                                     sin(Delta_rad) * cos(Hour_angle_rad))) * 180 / pi
+    sin(Delta_rad) * cos(Hour_angle_rad))) * 180 / pi
 
   return(Q_degr)
 }