+++ NEWS.md
@@ -5,10 +5,12 @@
 
 # specklestar v0.0.1.8 (2017-02-16)
 
-* Added par_angle.R function
+* Added par_angle.R and par_angle_manual.R functions for
+calculation of parallactic angles for BTA location
 
 # Roadmap of specklestar package
 
+- Remove log_date from par_angle.R
 - Function to find position of secondary maximum on ACF
 - Function to calculate contrast
 - Calculation of precise rho, phi and dm